requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.95.2
graphql-core==3.2.3
h11==0.14.0
httptools==0.5.0
idna==3.4
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
strawberry-graphql==0.177.1
typing_extensions==4.5.0
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
